worker/requirements.txt
# Celery
celery[redis]==5.4.0
flower==2.0.1

# Database
sqlalchemy==2.0.35
pymysql==1.1.1
cryptography==43.0.1

# Redis
redis==5.2.0

# MQTT
paho-mqtt==2.1.0

# HTTP Client
httpx==0.27.2

# File Processing
pillow==11.0.0
python-magic==0.4.27

# Utils
python-dotenv==1.0.1
